business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/rotator/Loc2LonLat.java
package eu.estcube.gsconnector.commands.rotator;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class Loc2LonLat implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+l ");
        messageString.append(command.getParameter("Locator"));
        messageString.append("\n");
        return messageString;
    }
}
